test: introduces an e2e test suite for operator
This is the beginning of an E2E test suite that uses Jobless to
provision test infrastructure and Kubernetes. It relies on a modified
version of Jobless that exports some internal functionality. This also
includes changes to the functional test files in order to re-use some
functionality there.

Status:

Provision or connect to an already provisioned cluster
Check status of cluster using conditions on CRD